+++ tauro/blueprints/api_oauth2_v1/endpoints/actualizar_turno_estado.py
@@ -138,14 +138,20 @@ class ActualizarTurnoEstado(Resource):
         socketio.send(one_turno_out)
 
         # Añadir al voceador el turno
-        if turno.turno_estado.nombre in ["PASE A VENTANILLA", "ATENDIENDO EN CUBICULO"]:
+        if turno.turno_estado.nombre in ["PASE A UBICACION", "PASE A CUBICULO"]:
             voceador_turnos = VocearTurnos()
             try:
                 resultado, mensaje_resp = voceador_turnos.agregar_mensaje(turno)
             except Exception as e:
                 pass
         # Quitar del voceador
-        elif turno.turno_estado.nombre in ["ATENDIENDO", "EN ESPERA DE CUBICULO", "CANCELADO", "COMPLETADO"]:
+        elif turno.turno_estado.nombre in [
+            "EN ESPERA DE CUBICULO",
+            "ATENDIENDO",
+            "ATENDIENDO EN CUBICULO",
+            "CANCELADO",
+            "COMPLETADO",
+        ]:
             voceador_turnos = VocearTurnos()
             try:
                 resultado, mensaje_resp = voceador_turnos.quitar_mensaje(turno)