+++ tauro/blueprints/api_oauth2_v1/endpoints/tomar_turno.py
@@ -84,16 +84,16 @@ class TomarTurno(Resource):
                 message="No hay turnos en espera",
             ).model_dump()
 
-        # Consultar el estado de turno "PASE A VENTANILLA"
+        # Consultar el estado de turno "PASE A UBICACION"
         try:
-            turno_estado = TurnoEstado.query.filter_by(nombre="PASE A VENTANILLA").one()
+            turno_estado = TurnoEstado.query.filter_by(nombre="PASE A UBICACION").one()
         except MultipleResultsFound, NoResultFound:
             return OneTurnoOut(
                 success=False,
                 message="Estado de turno no encontrado",
             ).model_dump()
 
-        # Cambiar el usuario, el estado a "PASE A VENTANILLA" y la ubicación, así como el tiempo de inicio
+        # Cambiar el usuario, el estado a "PASE A UBICACION" y la ubicación, así como el tiempo de inicio
         turno.usuario_id = usuario.id
         turno.turno_estado_id = turno_estado.id
         turno.ubicacion_id = usuario.ubicacion_id